requirements.txt
apify==3.1.0
playwright==1.49.1
rapidfuzz==3.10.1
unidecode==1.3.8
pyyaml==6.0.2
aiohttp==3.10.11
